% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDSskipInfoTagFloat}
\alias{getDSskipInfoTagFloat}
\title{get DS (dosage) of each sample and skip variants given a INFO tag of Float type and a cutoff}
\usage{
getDSskipInfoTagFloat(cutoff, tag, vcffile, region = "", samples = "-")
}
\arguments{
\item{cutoff}{skip variants with tag value smaller than this cutoff}

\item{tag}{tag exists in INFO column}

\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract}

\item{samples}{samples to extract}
}
\value{
A list of genotype dosages for each sample along with variant information in VCF
}
\description{
get DS (dosage) of each sample and skip variants given a INFO tag of Float type and a cutoff
}
