% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarySVs}
\alias{summarySVs}
\title{report the stats of structure variants}
\usage{
summarySVs(vcffile, region = "", samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract, default "" for all}

\item{samples}{samples to extract, default "-" for all}
}
\value{
the counts of each type of structure variant
}
\description{
report the stats of structure variants
}
