% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summaryVariants}
\alias{summaryVariants}
\title{report the stats of variants}
\usage{
summaryVariants(vcffile, region = "", samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract, default "" for all}

\item{samples}{samples to extract, default "-" for all}
}
\value{
the counts of each type of variant
}
\description{
report the stats of variants
}
