% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tableGP}
\alias{tableGP}
\title{parse GP FORMAT of the VCF file into tables in R}
\usage{
tableGP(vcffile, region, samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract}

\item{samples}{samples to extract}
}
\value{
A list of genotype posterior probabilites for each sample along with variant information in VCF
}
\description{
parse GP FORMAT of the VCF file into tables in R
}
