% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tableGT}
\alias{tableGT}
\title{parse GT of a VCF file into tables in R}
\usage{
tableGT(vcffile, region, samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract}

\item{samples}{samples to extract}
}
\value{
A list of genotypes and other fixed fields in VCF
}
\description{
parse GT of a VCF file into tables in R
}
