% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tableHQ}
\alias{tableHQ}
\title{parse HQ FORMAT of the VCF file into tables in R}
\usage{
tableHQ(vcffile, region, samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract}

\item{samples}{samples to extract}
}
\value{
A list of phred-scaled haplotype quality for each sample along with variant information in VCF
}
\description{
parse HQ FORMAT of the VCF file into tables in R
}
