% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tableMQ}
\alias{tableMQ}
\title{parse MQ FORMAT of the VCF file into tables in R}
\usage{
tableMQ(vcffile, region, samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract}

\item{samples}{samples to extract}
}
\value{
A list of RMS mapping quality for each sample along with variant information in VCF
}
\description{
parse MQ FORMAT of the VCF file into tables in R
}
