% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tablePL}
\alias{tablePL}
\title{parse PL FORMAT of the VCF file into tables in R}
\usage{
tablePL(vcffile, region, samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract}

\item{samples}{samples to extract}
}
\value{
A list of phred-scaled genotype likelihoods for each sample along with variant information in VCF
}
\description{
parse PL FORMAT of the VCF file into tables in R
}
