% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcfpp_calc_info_persite}
\alias{vcfpp_calc_info_persite}
\title{Calculate INFO score from GP after genotype imputation}
\usage{
vcfpp_calc_info_persite(GP)
}
\arguments{
\item{GP}{vector of length a multiple of 3}
}
\description{
Calculate INFO score from GP after genotype imputation
}
