% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfppR-package.R
\docType{package}
\name{vcfppR-package}
\alias{vcfppR}
\alias{vcfppR-package}
\title{vcfppR: Rapid Manipulation of the Variant Call Format (VCF) in R}
\description{
The \url{https://github.com/Zilong-Li/vcfpp} is an easy-to-use C++ API of htslib, offering the full functionalities as the htslib to manipulate the VCF/BCF file. Thus, this package is built upon the vcfpp.h to provide rapid variant processing with the variant call format in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zilong-Li/vcfppR}
  \item Report bugs at \url{https://github.com/Zilong-Li/vcfppR/issues}
}

}
\author{
\strong{Maintainer}: Zilong Li \email{zilong.dk@gmail.com} (\href{https://orcid.org/0000-0001-5859-2078}{ORCID}) [copyright holder]

}
