% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf-tables.R
\name{vcftable}
\alias{vcftable}
\title{read VCF/BCF contents into R data structure}
\usage{
vcftable(vcffile, region = "", samples = "-", format = "GT", ploidy = 2)
}
\arguments{
\item{vcffile}{path to the VCF/BCF file}

\item{region}{region to subset like bcftools}

\item{samples}{samples to subset like bcftools}

\item{format}{the FORMAT tag to extract, valid values are
"GT","GP", "DP","DS","GL","PL","GQ","HQ","MQ","PQ"}

\item{ploidy}{the ploidy of the organism if it is not diploidy}
}
\value{
\code{vcftable} a list containing the following components:
\describe{
\item{samples}{  character vector; \cr
the samples ids in the VCF file after subsetting
}

\item{chr}{  character vector; \cr
the CHR column in the VCF file
}

\item{pos}{  character vector; \cr
the POS column in the VCF file
}

\item{id}{  character vector; \cr
the ID column in the VCF file
}

\item{ref}{  character vector; \cr
the REF column in the VCF file
}

\item{alt}{  character vector; \cr
the ALT column in the VCF file
}

\item{qual}{  character vector; \cr
the QUAL column in the VCF file
}

\item{filter}{  character vector; \cr
the FILTER column in the VCF file
}

\item{info}{  character vector; \cr
the INFO column in the VCF file
}

\item{format}{  matrix of either integer of numberic values depending on the tag to extract; \cr
a specifiy tag in the FORMAT column to be extracted
}
}
}
\description{
If you want to read VCF/BCF into R data types fast and simplely, this is the swiss army knife
}
\details{
bcftools view -s "id01,id02" input.bcf.gz chr1:100000-20000
}
\examples{
library('vcfppR')
vcffile <- system.file("extdata", "test-GL.vcf.gz", package="vcfppR")
(res <- vcftable(vcffile,"chr20"))
vcffile <- system.file("extdata", "test-PL.vcf.gz", package="vcfppR")
(res <- vcftable(vcffile,"chr20", format = "PL"))
str(res)
}
\author{
Zilong Li \email{zilong.dk@gmail.com}
}
