% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcfwriter}
\alias{vcfwriter}
\title{API for writing the VCF/BCF.}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor given a vcf file \itemize{
\item Parameter: vcffile - The path of a vcf file. don't start with "~"
\item Parameter: version - The version of VCF specification
}}
}}

