% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heterozygosity}
\alias{heterozygosity}
\title{calculate the number of heterozygous SNPs for each sample}
\usage{
heterozygosity(vcffile, region = "", samples = "-")
}
\arguments{
\item{vcffile}{path to the VCF file with index}

\item{region}{region to extract, default "" for all}

\item{samples}{samples to extract, default "-" for all}
}
\value{
A list of heterozygosity couts for each sample along with its id in the vcf header
}
\description{
calculate the number of heterozygous SNPs for each sample
}
