% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf-compare.R
\name{vcfcomp}
\alias{vcfcomp}
\title{Compare two VCF/BCF files reporting various statistics}
\usage{
vcfcomp(
  test,
  truth,
  region = "",
  samples = "-",
  names = NULL,
  format = c("DS", "GT"),
  stats = "r2",
  bins = NULL,
  af = NULL,
  out = NULL,
  vartype = "snps",
  ids = NULL,
  qual = 0,
  pass = TRUE
)
}
\arguments{
\item{test}{path to the first VCF/BCF file referred as test.}

\item{truth}{path to the second VCF/BCF file referred as truth.}

\item{region}{region to subset in bcftools-like style: "chr1", "chr1:1-10000000"}

\item{samples}{samples to subset in bcftools-like style.
comma separated list of samples to include (or exclude with "^" prefix).
e.g. "id01,id02", "^id01,id02".}

\item{names}{rename samples in test VCF.}

\item{format}{character vector. the FORMAT tag to extract for comparison.
default c("DS", "GT") is used to extract DS of target and GT of truth respectively.}

\item{stats}{choose the statistics to be returned. eg. "r2", "f1"}

\item{bins}{numeric vector. allele frequency bins to stratify with.}

\item{af}{file path to allele frequency.}

\item{out}{output prefix for saving objects into RDS file}

\item{vartype}{restrict to specific type of variants. supports "snps","indels", "sv", "multisnps","multiallelics"}

\item{ids}{character vector. restrict to sites with ID in the given vector. default NULL won't filter any sites.}

\item{qual}{logical. restrict to variants with QUAL > qual.}

\item{pass}{logical. restrict to variants with FILTER = "PASS".}
}
\value{
return various statistics
}
\description{
Compare two VCF/BCF files reporting various statistics
}
\details{
\code{vcfcomp} implements various statisitcs to compare two VCF/BCF files,
e.g. report genotype concocrdance, correlation stratified by allele frequency.
}
\examples{
library('vcfppR')
test <- system.file("extdata", "raw.gt.vcf.gz", package="vcfppR")
truth <- system.file("extdata", "raw.gt.vcf.gz", package="vcfppR")
suppressWarnings(
  res <- vcfcomp(test, truth, stats = "f1", format = c("GT", "GT"))
)
str(res[!is.na(res)])
}
\author{
Zilong Li \email{zilong.dk@gmail.com}
}
