% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf-compare.R
\name{vcfcomp}
\alias{vcfcomp}
\title{Compare two VCF/BCF files reporting various statistics}
\usage{
vcfcomp(
  test,
  truth,
  formats = c("DS", "GT"),
  stats = "r2",
  bysample = FALSE,
  byvariant = FALSE,
  names = NULL,
  bins = NULL,
  af = NULL,
  out = NULL,
  ...
)
}
\arguments{
\item{test}{path to the first VCF/BCF file referred as test.}

\item{truth}{path to the second VCF/BCF file referred as truth, or saved RDS file.}

\item{formats}{character vector. the FORMAT tags to extract for the test and truth respectively.
default c("DS", "GT") extracts 'DS' of the target and 'GT' of the truth.}

\item{stats}{the statistics to be calculated. supports the following.
"r2": pearson correlation coefficient ** 2.
"f1": F1-score, good balance between sensitivity and precision.
"nrc": Non-Reference Concordance rate}

\item{bysample}{logical. calculate concordance for each samples, then average by bins.}

\item{byvariant}{logical. calculate concordance for each variant, then average by bins.
if both bysample and by variant are TRUE, then do average on all samples first.
if both bysample and by variant are FALSE, then do average on all samples and variants.}

\item{names}{character vector. reset samples' names in test VCF.}

\item{bins}{numeric vector. break statistics into allele frequency bins.}

\item{af}{file path to allele frequency text file or saved RDS file.}

\item{out}{output prefix for saving objects into RDS file}

\item{...}{options passed to \code{vcftable}}
}
\value{
a list of various statistics
}
\description{
Compare two VCF/BCF files reporting various statistics
}
\details{
\code{vcfcomp} implements various statisitcs to compare two VCF/BCF files,
e.g. report genotype concocrdance, correlation stratified by allele frequency.
}
\examples{
library('vcfppR')
test <- system.file("extdata", "imputed.gt.vcf.gz", package="vcfppR")
truth <- system.file("extdata", "imputed.gt.vcf.gz", package="vcfppR")
samples <- "HG00133,HG00143,HG00262"
res <- vcfcomp(test, truth, stats="f1", format=c('GT','GT'), samples=samples)
str(res)
}
\author{
Zilong Li \email{zilong.dk@gmail.com}
}
