% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf-plot.R
\name{vcfplot}
\alias{vcfplot}
\title{Do plotting based on various objects in vcfppR}
\usage{
vcfplot(obj, which.sample = 1, ...)
}
\arguments{
\item{obj}{object returned by vcfcomp}
}
\description{
Do plotting based on various objects in vcfppR
}
