% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf-plot.R
\name{vcfplot}
\alias{vcfplot}
\title{Make sensible and beautiful plots based on various objects in vcfppR}
\usage{
vcfplot(obj, which.sample = NULL, variant = c("SNP", "INDEL"), pop = NULL, ...)
}
\arguments{
\item{obj}{object returned by vcftable, vcfcomp, vcfsummary}

\item{which.sample}{which sample to be plotted. NULL will aggregate all samples.}

\item{variant}{which types of variant are desired}

\item{pop}{file contains population information}

\item{...}{parameters passed to graphics}
}
\description{
Make sensible and beautiful plots based on various objects in vcfppR
}
