% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf-plot.R
\name{vcfplot}
\alias{vcfplot}
\title{Make sensible and beautiful plots based on various objects in vcfppR}
\usage{
vcfplot(
  obj,
  what = "r2",
  which.sample = 1,
  variant = c("SNP", "INDEL"),
  pop = NULL,
  ...
)
}
\arguments{
\item{obj}{object returned by vcfcomp}

\item{what}{what statisitcs to be plotted}

\item{which.sample}{which sample among all to be plotted}

\item{variant}{which types of variant are desired}

\item{pop}{file contains population information}

\item{...}{parameters passed to graphics}
}
\description{
Make sensible and beautiful plots based on various objects in vcfppR
}
