% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfppR-package.R
\docType{package}
\name{vcfppR-package}
\alias{vcfppR}
\alias{vcfppR-package}
\title{vcfppR: Rapid Manipulation of the Variant Call Format (VCF)}
\description{
The \url{https://github.com/Zilong-Li/vcfpp} is an easy-to-use 'C++' 'API' of 'htslib', offering the full functionalities as the 'htslib' to manipulate the variant call format (VCF) file. Thus, this package is built upon the 'vcfpp.h' for rapid variant processing of the compressed or uncompressed VCF/BCF file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zilong-Li/vcfppR}
  \item Report bugs at \url{https://github.com/Zilong-Li/vcfppR/issues}
}

}
\author{
\strong{Maintainer}: Zilong Li \email{zilong.dk@gmail.com} (\href{https://orcid.org/0000-0001-5859-2078}{ORCID})

Other contributors:
\itemize{
  \item Bonfield, James K and Marshall, John and Danecek, Petr and Li, Heng (Authors of included htslib library) [copyright holder]
}

}
