% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfppR-package.R
\docType{package}
\name{vcfppR-package}
\alias{vcfppR}
\alias{vcfppR-package}
\title{Imports}
\description{
This package demonstrates the usage of vcfpp in R
}
\author{
\strong{Maintainer}: Zilong Li \email{zilong.dk@gmail.com} (\href{https://orcid.org/0000-0001-5859-2078}{ORCID})

}
