% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcfreader}
\alias{vcfreader}
\title{API for reading read the VCF/BCF.}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor given a vcf file \itemize{
\item Parameter: vcffile - The path of a vcf file
}}

\item{\code{new}}{Constructor given a vcf file and the region \itemize{
\item Parameter: vcffile - The path of a vcf file
\item Parameter: region - The region to be constrained
}}

\item{\code{new}}{Constructor given a vcf file, the region and the samples \itemize{
\item Parameter: vcffile - The path of a vcf file
\item Parameter: region - The region to be constrained
\item Parameter: samples - The samples to be constrained. Comma separated list of samples to include (or exclude with "^" prefix).
}}

\item{\code{variant}}{Try to get next variant record. Return false if there are no more variants or hit the end of file, otherwise return true.}

\item{\code{chr}}{Return the CHROM field of current variant}

\item{\code{pos}}{Return the POS field of current variant}

\item{\code{id}}{Return the CHROM field of current variant}

\item{\code{ref}}{Return the REF field of current variant}

\item{\code{alt}}{Return the ALT field of current variant}

\item{\code{qual}}{Return the QUAL field of current variant}

\item{\code{filter}}{Return the FILTER field of current variant}

\item{\code{info}}{Return the INFO field of current variant}

\item{\code{infoInt}}{Return the tag value of integer type in INFO field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in INFO}}

\item{\code{infoFloat}}{Return the tag value of float type in INFO field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in INFO}}

\item{\code{infoStr}}{Return the tag value of string type in INFO field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in INFO}}

\item{\code{infoIntVec}}{Return the tag value in a vector of integer type in INFO field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in INFO}}

\item{\code{infoFloatVec}}{Return the tag value in a vector of float type in INFO field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in INFO}}

\item{\code{genotypes}}{Return the genotype values in a vector of integers  \itemize{ \item Parameter: collapse - Boolean value indicates wheather to collapse the size of genotypes, eg, return diploid genotypes.}}

\item{\code{formatInt}}{Return the tag value of integer type for each sample in FORAMT field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in FORAMT}}

\item{\code{formatFloat}}{Return the tag value of float type for each sample in FORAMT field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in FORAMT}}

\item{\code{formatStr}}{Return the tag value of string type for each sample in FORAMT field of current variant \itemize{ \item Parameter: tag - The tag name to retrieve in FORAMT}}

\item{\code{nsamples}}{Return the number of samples}

\item{\code{isSNP}}{Test if current variant is exculsively a SNP or not}

\item{\code{isIndel}}{Test if current variant is exculsively a INDEL or not}

\item{\code{isSV}}{Test if current variant is exculsively a SV or not}

\item{\code{isMultiAllelics}}{Test if current variant is exculsively a Multi Allelics or not}

\item{\code{isMultiAllelicSNP}}{Test if current variant is exculsively a Multi Biallelics (SNPs) or not}

\item{\code{hasSNP}}{Test if current variant has a SNP or not}

\item{\code{hasINDEL}}{Test if current variant has a INDEL or not}

\item{\code{hasINS}}{Test if current variant has a INS or not}

\item{\code{hasDEL}}{Test if current variant has a DEL or not}

\item{\code{hasMNP}}{Test if current variant has a MNP or not}

\item{\code{hasBND}}{Test if current variant has a BND or not}

\item{\code{hasOTHER}}{Test if current variant has a OTHER or not}

\item{\code{hasOVERLAP}}{Test if current variant has a OVERLAP or not}
}}

