% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcfwriter}
\alias{vcfwriter}
\title{API for writing the VCF/BCF.}
\value{
A class with many methods for outputting the VCF/BCF
}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor given a vcf file \itemize{
\item Parameter: vcffile - The path of a vcf file. don't start with "~"
\item Parameter: version - The version of VCF specification
}}

\item{\code{addContig}}{Add a Contig in the header of the vcf}

\item{\code{addFILTER}}{Add a FILTER in the header of the vcf}

\item{\code{addINFO}}{Add a INFO in the header of the vcf}

\item{\code{addFORMAT}}{Add a FORMAT in the header of the vcf}

\item{\code{addSample}}{Add a SAMPLE in the header of the vcf}

\item{\code{addLine}}{Add a line in the header of the vcf}

\item{\code{writeline}}{Write a variant record given a line}

\item{\code{close}}{Close and save the vcf file}
}}

\examples{
outvcf <- paste0(tempfile(), ".vcf.gz")
bw <- vcfwriter$new(outvcf, "VCF4.3")
bw$addContig("chr20")
bw$addINFO("AF", "A", "Float", "Estimated allele frequency in the range (0,1)");
bw$addFORMAT("GT", "1", "String", "Genotype");
bw$addSample("NA12878")
s1 <- "chr20\t2006060\trs146931526\tG\tC\t100\tPASS\tAF=0.000998403\tGT\t1|0"
bw$writeline(s1)
bw$close()
}
